move the prettyprinting stuff around; fewer orphans, trace functions
