+++ core/src/Language/Avaleryar/PDP.hs
@@ -127,7 +127,7 @@ runQuery :: Monad m => [Fact] -> Text -> [Term TextVar] -> PDP m [Fact]
 runQuery facts p args  = do
   answers <- runAvaWith (insertApplicationAssertion facts) $ compileQuery "system" p args
   flip traverse answers $ \l -> do
-     traverse (throwError . VarInQueryResults . snd) l
+     traverse (throwError . VarInQueryResults . unEVar) l
 
 runQuery' :: MonadIO m => [Fact] -> Query -> PDP m [Fact]
 runQuery' facts (Lit (Pred p _) as) = runQuery facts p as
