+++ core/src/Language/Avaleryar/Syntax.hs
@@ -4,8 +4,10 @@
 {-# LANGUAGE DeriveTraversable          #-}
 {-# LANGUAGE FlexibleInstances          #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE OverloadedStrings          #-}
 {-# LANGUAGE PolyKinds                  #-}
 {-# LANGUAGE StandaloneDeriving         #-}
+{-# LANGUAGE TypeApplications           #-}
 {-# LANGUAGE TypeSynonymInstances       #-}
 
 {-|
@@ -41,11 +43,15 @@ name of the assertion in which @may\/1@ is defined).  In brief:
 
 module Language.Avaleryar.Syntax where
 
-import Data.Functor.Const (Const(..))
-import Data.Map           (Map)
-import Data.String
-import Data.Text          (Text)
-import Data.Void
+import           Data.Char                    (isSpace)
+import           Data.Functor.Const           (Const(..))
+import           Data.Map                     (Map)
+import           Data.String
+import           Data.Text                    (Text)
+import qualified Data.Text                    as T
+import           Data.Void
+import           Text.PrettyPrint.Leijen.Text
+    (Doc, Pretty(..), brackets, colon, dot, empty, group, hsep, line, nest, parens, punctuate, space, vsep)
 
 data Value
   = I Int
@@ -56,8 +62,18 @@ data Value
 instance IsString Value where
   fromString = T . fromString
 
+instance Pretty Value where
+  pretty (I n) = pretty n
+  pretty (B b) = if b then "#t" else "#f"
+  pretty (T t) = if T.any isSpace t
+                 then pretty (show t) -- want the quotes/escaping
+                 else pretty t        -- display as a symbol
+
 -- | A predicate is identified by its name and arity (i.e., the predicate of the literal @foo(bar, ?baz)@ is @foo/2@)
 data Pred = Pred Text Int deriving (Eq, Ord, Read, Show)
+
+instance Pretty Pred where
+  pretty (Pred p n) = pretty p <> "/" <> pretty n
 
 -- | A term is either a 'Value' or a variable.  Terms are polymorphic in the variable type to
 -- provide a bit of safety by keeping us from crossing various streams (e.g., separating runtime
@@ -65,9 +81,16 @@ data Pred = Pred Text Int deriving (Eq, Ord, Read, Show)
 -- variable capture).
 data Term v = Val Value | Var v deriving (Eq, Ord, Read, Show, Functor, Foldable, Traversable)
 
+instance Pretty v => Pretty (Term v) where
+  pretty (Var v) = "?" <> pretty v
+  pretty (Val c) = pretty c
+
 -- | A literal is identified by a 'Pred' and a list of 'Term's, where the arity in the 'Pred' is the
 -- same as the length of the list of 'Term's in the argument list.
 data Lit v = Lit Pred [Term v] deriving (Eq, Ord, Read, Show, Functor, Foldable, Traversable)
+
+instance Pretty v => Pretty (Lit v) where
+  pretty (Lit (Pred p _) as) = pretty p <> parens (hsep . punctuate "," $ fmap pretty as)
 
 -- | Convenience constructor for 'Lit's.
 lit :: Text -> [Term v] -> Lit v
@@ -77,22 +100,40 @@ lit pn as = Lit (Pred pn (length as)) as
 -- 'Term'.
 data ARef v = ARNative Text | ARTerm (Term v) deriving (Eq, Ord, Read, Show, Functor, Foldable, Traversable)
 
+instance Pretty v => Pretty (ARef v) where
+  pretty (ARTerm t)   = pretty t
+  pretty (ARNative n) = colon <> pretty n
+
+prettyAssertion :: Value -> [Pred] -> Doc
+prettyAssertion assn ps = pretty assn
+                       <> ": "
+                       <> group (nest 2 (line <> (vsep . fmap pretty $ ps)))
+
 -- | A 'Lit'eral appearing in the body of a 'Rule' is always qualified by an 'ARef' to an assertion.
 -- When no assertion appears in the concrete syntax, the parser inserts a reference to the assertion
 -- currently being parsed.
 data BodyLit v = Says (ARef v) (Lit v)
   deriving (Eq, Ord, Read, Show, Functor, Foldable, Traversable)
 
+instance Pretty v => Pretty (BodyLit v) where
+  pretty (aref `Says` l) = pretty aref <> space <> "says" <> space <> pretty l
+
 -- | A rule has a head and a body made of 'BodyLit's.
 data Rule v = Rule (Lit v) [BodyLit v]
   deriving (Eq, Ord, Read, Show, Functor, Foldable, Traversable)
+
+instance Pretty v => Pretty (Rule v) where
+  pretty (Rule hd body) = pretty hd <> bodyDoc body <> dot <> line
+    where bodyDoc [] = empty
+          bodyDoc _  = space <> ":-"
+                    <> group (nest 2 (line <> (vsep . punctuate "," $ fmap pretty body)))
 
 -- | Facts can be thought of as rules with no variables in the head and no body.  Instead, we
 -- represent them as 'Lit's with variables of type 'Void' to ensure they are facts by construction.
 type Fact = Lit Void
 
--- | Convenience constructor for 'Fact's.  Works directly only 'Valuable's because the whole point
--- is that there aren't any variables amongst the arguments.
+-- | Convenience constructor for 'Fact's.  Works directly on 'Valuable's because the whole point is
+-- that there aren't any variables amongst the arguments.
 fact :: Valuable v => Text -> [v] -> Fact
 fact pn = lit pn . fmap val
 
@@ -106,6 +147,12 @@ factToRule fct = Rule (vacuous fct) []
 -- manipulate code /before/ it's loaded into the system and evaluated.  The motivating use cases for
 -- directives are to declare test-suites and (eventually) mode declarations.
 data Directive = Directive Fact [Fact] deriving (Eq, Ord, Show)
+
+-- TODO: Pretty this up.
+instance Pretty Directive where
+  pretty (Directive f fs) = ":- " <> pretty (vacuous @_ @TextVar f)
+                                  <> group (nest 2 (line <> (vsep . fmap (pretty . vacuous @_ @TextVar) $ fs)))
+                                  <> dot
 
 -- | To ensure freshness, tag runtime variables ('EVar's) with the current value of an 'Epoch'
 -- counter which we bump every time we allocate a new variable.
@@ -122,13 +169,21 @@ query :: Text -> [Term TextVar] -> Query
 query = lit
 
 -- | At runtime, we unify on variables tagged with an 'Epoch' to help avoid undesirable variable capture.
---
--- TODO: This should probably also become a newtype
-type EVar = (Epoch, TextVar)
+data EVar = EVar Epoch TextVar deriving (Eq, Ord, Read, Show)
+
+instance Pretty EVar where
+  pretty (EVar (Epoch e) v) = pretty v <> brackets (pretty e)
+
+-- | Extract the 'TextVar' portion of an 'EVar'.
+unEVar :: EVar -> TextVar
+unEVar (EVar _ v) = v
 
 -- | Raw variables are produced by the parser.
 newtype RawVar = RawVar { unRawVar :: Text }
   deriving (Eq, Ord, Read, Show, IsString)
+
+instance Pretty RawVar where
+  pretty = pretty . unRawVar
 
 -- | The runtime substitution.
 type Env = Map EVar (Term EVar)
